// clientsrc/src/app/ims/fields/field-base/field-type.ts
export enum FieldType {
    Text = 1,
    Numeric = 2,
    RichText = 3,
    Tags = 4,
    PredefinedList = 5,
    EntityReference = 6,
    Checkbox = 7
}

export const FieldTypes = {
    'Text': FieldType.Text,
    'Numeric': FieldType.Numeric,
    'RichText': FieldType.RichText,
    'Tags': FieldType.Tags,
    'Predefined List': FieldType.PredefinedList,
    'Entity Reference': FieldType.EntityReference,
    'Checkbox': FieldType.Checkbox
};